Install tools.py and move Solver class into a separate file.
